requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
